Move parsers to their own package.
